Move `Services` api to `.service._mngr` mod
